.history/数据库详情/review_system架构（评价表）/测试/requirements_20241220170734.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
numpy>=1.24.0
python-dotenv>=1.0.0 